move caching extension to relative class
